requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
kinde-python-sdk==1.2.3
pytest==8.0.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
pytest-cov==4.1.0
starlette==0.36.3
pydantic==2.6.1
pydantic-settings==2.1.0
